# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Samryetha API",
    "version": "0.1.5"
  },
  "paths": {
...
    "/api/discussions": {
      "get": {
        "summary": "List Discussions",
        "operationId": "list_discussions_api_discussions_get",
        "parameters": [
          {
            "name": "feed",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "latest",
                "followed"
              ],
              "type": "string",
              "default": "latest",
              "title": "Feed"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "date",
                "replies"
              ],
              "type": "string",
              "default": "date",
              "title": "Sort"
            }
          },
          {
            "name": "board",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Board"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Discussions Api Discussions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Discussion",
        "operationId": "create_discussion_api_discussions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiscussionBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Discussion Api Discussions Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/discussions/preview": {
      "post": {
        "summary": "Preview Body",
        "operationId": "preview_body_api_discussions_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Preview Body Api Discussions Preview Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/discussions/{discussion_id}": {
      "get": {
        "summary": "Get Discussion",
        "operationId": "get_discussion_api_discussions__discussion_id__get",
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Discussion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Discussion Api Discussions  Discussion Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Discussion",
        "operationId": "update_discussion_api_discussions__discussion_id__patch",
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Discussion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDiscussionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Discussion Api Discussions  Discussion Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Discussion",
        "operationId": "delete_discussion_api_discussions__discussion_id__delete",
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Discussion Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DeleteDiscussionBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Discussion Api Discussions  Discussion Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/discussions/{discussion_id}/replies": {
      "post": {
        "summary": "Create Reply",
        "operationId": "create_reply_api_discussions__discussion_id__replies_post",
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Discussion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReplyBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Reply Api Discussions  Discussion Id  Replies Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Replies",
        "operationId": "list_replies_api_discussions__discussion_id__replies_get",
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Discussion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Replies Api Discussions  Discussion Id  Replies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/replies/{reply_id}": {
      "patch": {
        "summary": "Update Reply",
        "operationId": "update_reply_api_replies__reply_id__patch",
        "parameters": [
          {
            "name": "reply_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Reply Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReplyBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Reply Api Replies  Reply Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Reply",
        "operationId": "delete_reply_api_replies__reply_id__delete",
        "parameters": [
          {
            "name": "reply_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Reply Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Reply Api Replies  Reply Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/discussions/{discussion_id}/save": {
      "post": {
        "summary": "Save Discussion",
        "operationId": "save_discussion_api_discussions__discussion_id__save_post",
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Discussion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Save Discussion Api Discussions  Discussion Id  Save Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Unsave Discussion",
        "operationId": "unsave_discussion_api_discussions__discussion_id__save_delete",
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Discussion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Unsave Discussion Api Discussions  Discussion Id  Save Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/discussions/{discussion_id}/follow": {
      "post": {
        "summary": "Follow Discussion",
        "operationId": "follow_discussion_api_discussions__discussion_id__follow_post",
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Discussion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Follow Discussion Api Discussions  Discussion Id  Follow Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Unfollow Discussion",
        "operationId": "unfollow_discussion_api_discussions__discussion_id__follow_delete",
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Discussion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Unfollow Discussion Api Discussions  Discussion Id  Follow Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/discussions/{discussion_id}/pin": {
      "post": {
        "summary": "Pin Discussion",
        "operationId": "pin_discussion_api_discussions__discussion_id__pin_post",
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Discussion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Pin Discussion Api Discussions  Discussion Id  Pin Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/discussions/{discussion_id}/lock": {
      "post": {
        "summary": "Lock Discussion",
        "operationId": "lock_discussion_api_discussions__discussion_id__lock_post",
        "parameters": [
          {
            "name": "discussion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Discussion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Lock Discussion Api Discussions  Discussion Id  Lock Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDiscussionBody": {
        "properties": {
          "boardSlug": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Boardslug"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "bodyMarkdown": {
            "type": "string",
            "maxLength": 40000,
            "minLength": 1,
            "title": "Bodymarkdown"
          },
          "bodyFormat": {
            "type": "string",
            "enum": [
              "markdown",
              "text"
            ],
            "title": "Bodyformat",
            "default": "markdown"
          },
          "attachmentIds": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array",
                "maxItems": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachmentids"
          },
          "draftId": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Draftid"
          }
        },
        "type": "object",
        "required": [
          "boardSlug",
          "bodyMarkdown"
        ],
        "title": "CreateDiscussionBody"
      },
      "CreateReplyBody": {
        "properties": {
          "bodyMarkdown": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Bodymarkdown"
          },
          "bodyFormat": {
            "type": "string",
            "enum": [
              "markdown",
              "text"
            ],
            "title": "Bodyformat",
            "default": "markdown"
          },
          "parentReplyId": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentreplyid"
          }
        },
        "type": "object",
        "required": [
          "bodyMarkdown"
        ],
        "title": "CreateReplyBody"
      },
...
      "DeleteDiscussionBody": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "DeleteDiscussionBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreviewBody": {
        "properties": {
          "bodyMarkdown": {
            "type": "string",
            "maxLength": 40000,
            "title": "Bodymarkdown"
          },
          "bodyFormat": {
            "type": "string",
            "enum": [
              "markdown",
              "text"
            ],
            "title": "Bodyformat",
            "default": "markdown"
          }
        },
        "type": "object",
        "required": [
          "bodyMarkdown"
        ],
        "title": "PreviewBody"
      },
...
      "UpdateDiscussionBody": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "bodyMarkdown": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodymarkdown"
          },
          "bodyFormat": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "markdown",
                  "text"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodyformat"
          }
        },
        "type": "object",
        "title": "UpdateDiscussionBody"
      },
      "UpdateReplyBody": {
        "properties": {
          "bodyMarkdown": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Bodymarkdown"
          },
          "bodyFormat": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "markdown",
                  "text"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodyformat"
          }
        },
        "type": "object",
        "required": [
          "bodyMarkdown"
        ],
        "title": "UpdateReplyBody"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}